refactor(commands): View Parameters ➜ View Prompt
This change renames the "View Parameters" context menu command to "View Prompt" for clarity. The rationale behind this change is that while parameters may be more apt, most (especially the less technically-inclined) users are more familiar with the "Prompt" terminology, even if a prompt is only one portion of the puzzle for an image. 

This change is directed toward UX and Discoverability/a11y